Bug 785427 - GIMP crashes after applying any GEGL filter
Need to access GimpSettings' "time" property using gint64 variables
now.

+++ app/tools/gimpfiltertool.c
@@ -894,7 +894,9 @@ gimp_filter_tool_real_set_config (GimpFilterTool *filter_tool,
    *  will be considered to be among the automatically stored recently
    *  used settings
    */
-  g_object_set (filter_tool->config, "time", 0, NULL);
+  g_object_set (filter_tool->config,
+                "time", (gint64) 0,
+                NULL);
 }
 
 static void
